runtime: buffer factories use size_t
Addresses the issue where large buffer size requests get converted back
and forth to signed/unsigned and can cause an overflow and the buffer
allocation to fail

+++ gnuradio-runtime/lib/vmcircbuf.h
@@ -72,7 +72,7 @@ public:
      *
      * Call this to create a doubly mapped circular buffer.
      */
-    virtual vmcircbuf* make(int size) = 0;
+    virtual vmcircbuf* make(size_t size) = 0;
 };
 
 /*
@@ -90,7 +90,7 @@ public:
     static vmcircbuf_factory* get_default_factory();
 
     static int granularity() { return get_default_factory()->granularity(); }
-    static vmcircbuf* make(int size) { return get_default_factory()->make(size); }
+    static vmcircbuf* make(size_t size) { return get_default_factory()->make(size); }
 
     // N.B. not all factories are guaranteed to work.
     // It's too hard to check everything at config time, so we check at runtime
